// kservices_ts_api/src/api/domain/entities/v3/sme.entity.ts
export type callListFetchRequest = {
  id: number;
  type: string;
  isDownload: string;
  initialRecord: number;
  batchSize: number;
  startDate: string;
  startDate_op: string;
  endDate: string;
  endDate_op: string;
  callDirectionStatus: string;
  callDirectionStatus_op: string;
  duration: string;
  duration_op: string;
  calledNumber: string;
  calledNumber_op: string;
  callingNumber: string;
  callingNumber_op: string;
  agentName: string;
  agentName_op: string;
  callStatus: string;
  callStatus_op: string;
  answerStatus: string;
  answerStatus_op: string;
  remarks: string;
  remarks_op: string;
  callId: string;
  callId_op: string;
  callFlow: string,
  callFlow_op: string,
};
// ...
export type mergeCallFetchRequest = {
  smeId: number;
  initialRecord: number;
  batchSize: number;
  startDate: string;
  endDate: string;
  callDirection: string;
  duration: string;
  customerNumber: Object;
  customerName: string;
  agentNumber: string;
  agentName: string;
  callStatus: string;
  remarks: string;
  callFlow: string,
  searchLeads: string,
  searchLeads_category: string,
  callQueue: string,
};